1st Lab 1st Commit
Problem 1, 2 and 3 are already done by me.
2 and 3 was done by stack using structures, so I redid it by stack using arrays.
And 4th question is new and interesting. I had to sort it by taking it out into an array and then applying bubble sort. It feels like I cheated on it.
The 5th question requires sorting and I don't want to use array sorting because that is not the concept that matters. So I will attempt it and work on it later in the week.