Hanoi in Class
We solved the Tower of Hanoi problem in class and we were supposed to things a little differently.
We were asked to print the different stacks that were interchanged, so this one is a little different from the code that I had made by myself. So have a look at this one as well.